+++ pinot-tools/src/main/java/org/apache/pinot/tools/AuthZkBasicQuickstart.java
@@ -23,7 +23,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import org.apache.pinot.common.auth.AuthProviderUtils;
-import org.apache.pinot.core.auth.BasicAuthUtils;
+import org.apache.pinot.common.auth.BasicAuthUtils;
 import org.apache.pinot.spi.auth.AuthProvider;
 import org.apache.pinot.spi.plugin.PluginManager;
